可以添加彩票类型

+++ shopping/src/main/java/org/ql/shopping/controller/user/UserServiceMangerController.java
@@ -67,7 +67,7 @@ public class UserServiceMangerController {
 	 * @return
 	 */
 	@RequestMapping("/view/change")
-	public ModelAndView userServiceChange(@RequestParam("id")long id) {
+	public ModelAndView userServiceChange(@RequestParam("id")int id) {
 		ModelAndView mav = new ModelAndView();
 		UserManager changeUser = mUserManagerService.findUserById(id);
 		if(changeUser == null){
